### 2023-12-09 #6130 @editorialbot
[PRE REVIEW]: StreamPoseML: An End-to-End Open-Source Web Application and Python Toolkit  for Real-Time Video Pose Classification and Machine Learning
**Submitting author:** @mrilikecoding (<a href="http://orcid.org/0000-0003-0157-7744">N Green</a>)
**Repository:** https://github.com/mrilikecoding/StreamPoseML
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.3
