### 2024-01-20 #6130 @arfon
@mrilikecoding – thanks for your patience. I'm going to be the handling editor for this submission. Before we get started, could you please take a look at the error generating the paper (https://github.com/openjournals/joss-reviews/issues/6130#issuecomment-1848643516), and fix up your `paper.md`?

Also, as I start to look for possible reviewers, please share with me here any people that you think might be able to give this submission a good review.
