### 2024-01-20 906b894
update paper affiliation formatting
