+++ paper.md
@@ -1,5 +1,7 @@
 ---
-title: “StreamPoseML” - An End-to-End Open-Source Web Application and Python Toolkit  for Real-Time Video Pose Classification and Machine Learning
+---
+title: “StreamPoseML” - An End-to-End Open-Source Web Application and Python
+  Toolkit  for Real-Time Video Pose Classification and Machine Learning
 tags:
   - Movement AI classification
   - Media Pipe
@@ -8,24 +10,27 @@ authors:
   - name: Milka Trajkova
     orcid: 0000-0002-3694-763X
     equal-contrib: true
-    affiliation: 1 
+    affiliation: 1
   - name: Nathaniel Green
     orcid: 0000-0003-0157-7744
-    equal-contrib: true 
+    equal-contrib: true
     affiliation: 2
   - name: Minoru Shinohara
     orcid: 0000-0002-9260-0366
-    corresponding: true 
+    corresponding: true
     affiliation: “3, 4”
 affiliations:
- - name: School of Literature, Media, and Communication, Georgia Institute of Technology, Atlanta, GA, USA
-  index: 1
- - name: Independent Researcher, USA
-   index: 2
- - name: School of Biological Sciences, Georgia Institute of Technology, Atlanta, GA, USA
-  index: 3
- - name: Wallace H. Coulter Department of School of Biomedical Engineering, Georgia Institute of Technology and Emory University, Atlanta, GA, USA
-  index: 4
+  - name: School of Literature, Media, and Communication, Georgia Institute of
+      Technology, Atlanta, GA, USA
+    index: 1
+  - name: Independent Researcher, USA
+    index: 2
+  - name: School of Biological Sciences, Georgia Institute of Technology, Atlanta,
+      GA, USA
+    index: 3
+  - name: Wallace H. Coulter Department of School of Biomedical Engineering, Georgia
+      Institute of Technology and Emory University, Atlanta, GA, USA
+    index: 4
 date: 1 December 2023
 bibliography: paper.bib
 ---
