### 2024-01-31 #6130 @arfon
@SwathiSheshadri @MLHale @nulberry @abhishektha @tgeijten – :wave: would any of you be willing to review this submission for JOSS? The submission under consideration is _StreamPoseML: An End-to-End Open-Source Web Application and Python Toolkit for Real-Time Video Pose Classification and Machine Learning_

Based on your experience, and past submissions as authors in JOSS, we think you might be able to provide a great review of this submission. Please let me know if you think you can help us out!

Many thanks
Arfon
