### 2024-02-19 #6130 @arfon
@aaa34169 @imcatta @thejasvibr – 👋 would any of you be willing to review this submission for JOSS? The submission under consideration is _StreamPoseML: An End-to-End Open-Source Web Application and Python Toolkit for Real-Time Video Pose Classification and Machine Learning_

Many thanks
Arfon
