+++ paper.md
@@ -32,7 +32,7 @@ bibliography: paper.bib
 
 # Summary
 
-With the ubiquity of video content in various sectors, from surveillance to entertainment to education, the demand for effective video analytics has never been greater - video data are being produced and consumed at an unprecedented rate, leading to an increased need for end-to-end research tools to integrate, manage, and utilize this massive influx of information. Classifying videos using machine learning often involves a series of data processing steps – data collection, annotation, pre-processing, model training, and deployment. These tasks require different tools and skill sets, making the process fragmented, time-consuming, and complex. Currently, many open-source machine learning pipelines require users to manually integrate different tools and libraries, which can be time-consuming and error-prone. There is a significant need for tools that can automatically develop an end-to-end machine learning pipeline, especially for human motion classification. A unified platform that can handle multiple stages in the machine learning pipeline will greatly streamline the workflow, saving time, reducing complexity, and lowering the barrier to entry for practitioners without extensive machine learning expertise. Our proposed web application web application and Python toolkit, “StreamPoseML” is an open-source, end-to-end toolkit for creating real-time, video-based classification experiments based on labeled data and body keypoint/pose estimation of human movements. It provides the capability of dataset generation, pre-processing, model building and training, as well as deployment in a real-time environment. Users will be able to monitor and adjust their models on the fly in the backend and test their models using a front-end UI. To illustrate its use, we provide an examples of on-going research on designing a real-time trial-to-trial movement classification for older adults undergoing Adapted Tango dance therapy.
+With the ubiquity of video content in various sectors, from surveillance to entertainment to education, the demand for effective video analytics has never been greater - video data are being produced and consumed at an unprecedented rate, leading to an increased need for end-to-end research tools to integrate, manage, and utilize this massive influx of information. Classifying videos using machine learning often involves a series of data processing steps – data collection, annotation, pre-processing, model training, and deployment. These tasks require different tools and skill sets, making the process fragmented, time-consuming, and complex. Currently, many open-source machine learning pipelines require users to manually integrate different tools and libraries, which can be time-consuming and error-prone. There is a significant need for tools that can automatically develop an end-to-end machine learning pipeline, especially for human motion classification. A unified platform that can handle multiple stages in the machine learning pipeline will greatly streamline the workflow, saving time, reducing complexity, and lowering the barrier to entry for practitioners without extensive machine learning expertise. Our proposed web application and Python toolkit, “StreamPoseML” is an open-source, end-to-end toolkit for creating real-time, video-based classification experiments based on labeled data and body keypoint/pose estimation of human movements. It provides the capability of dataset generation, pre-processing, model building and training, as well as deployment in a real-time environment. Users will be able to monitor and adjust their models on the fly in the backend and test their models using a front-end UI. To illustrate its use, we provide an examples of on-going research on designing a real-time trial-to-trial movement classification for older adults undergoing Adapted Tango dance therapy.
 
 # Statement of Need
 
