### 2024-03-10 mrilikecoding/StreamPoseML#14 @imcatta
JOSS paper
Hi! I'm one of the reviewers of your JOSS submission openjournals/joss-reviews#6392
This PR is intended to keep track of things you might want to revise in the paper.

1. While the pipeline shown in Figure 1 is very interesting and common to various applications of ML pose estimation, I think some elements of the last two steps are very context specific, resulting in a restriction of the otherwise broad generability of the workflow. In particular, they are: (a) inference on videos captured via the web, since one can classify videos from other sources; (b) sending results via Bluetooth.
3. [Line 34] Missing reference to the figure.
4. [Line 44] A reference to Google's Teachable Machine should be added.
5. [Figure 1] The text in the figure is small compared to the font size of the rest of the manuscript.
