### 2024-03-15 a7932ad
pull bluetooth into isolated component; bump mediapipe and dep load
