+++ web_ui/src/VideoStream.js
@@ -4,7 +4,8 @@ import {
     PoseLandmarker,
     FilesetResolver,
     // DrawingUtils
-} from "https://cdn.skypack.dev/@mediapipe/tasks-vision@0.10.0";
+} from "@mediapipe/tasks-vision";
+
 
 let poseLandmarker = undefined;
 const createPoseLandmarker = async () => {
@@ -23,16 +24,9 @@ const createPoseLandmarker = async () => {
 // TODO move this flag into UI - here for easy testing right now
 const USE_CLIENTSIDE_POSE_ESTIMATION = true;
 
-// Bluetooth - TODO abstract this out so that we can select what actuator to use
-const DEVICE_SERVICE_UUID = process.env.REACT_APP_BLUETOOTH_DEVICE_SERVICE_UUID.toLowerCase();
-const DEVICE_CHARACTERISTIC_UUID = process.env.REACT_APP_BLUETOOTH_DEVICE_CHARACTERISTIC_UUID.toLowerCase();
-
-function VideoStream({ isOn = false }) {
+function VideoStream({ isOn = false, handleClassification }) {
     const localVideoRef = useRef();
     const [results, setResults] = useState(null);
-    const [characteristic, setCharacteristic] = useState(null);
-    const [bluetoothStatus, setBluetoothStatus] = useState("Connect to Bluetooth");
-    const [bluetoothResult, setBluetoothResult] = useState(null);
     const socketRef = useRef();
 
     useEffect(() => {
@@ -88,33 +82,10 @@ function VideoStream({ isOn = false }) {
             initWebRTC();
 
             // Set up the Socket.IO connection and event listeners
-            socketRef.current = io.connect("http://localhost:5001");
+            socketRef.current = io.connect(process.env.REACT_APP_STREAM_POSE_ML_API_ENDPOINT);
             socketRef.current.on("frame_result", (data) => {
                 setResults(data);
-                if (data["classification"] === true && characteristic) {
-                    // send time in console
-                    let encoder = new TextEncoder('utf-8');
-                    let value = encoder.encode('a');
-                    characteristic.writeValue(value)
-                        .then(() => {
-                            console.log('Write operation is complete.');
-                            // Now read from the characteristic
-                            return characteristic.readValue();
-                        })
-                        .then(value => {
-                            let decoder = new TextDecoder('utf-8');
-                            let result = decoder.decode(value);
-                            console.log('Read operation result:', result);
-                            setBluetoothResult(result);
-                        })
-                        .catch(error => {
-                            console.log(error)
-                            setBluetoothResult("")
-                        });
-                } else {
-                    setBluetoothResult("")
-                }
-
+                handleClassification(data)
             });
 
 
@@ -193,61 +164,17 @@ function VideoStream({ isOn = false }) {
                 socketRef.current.disconnect();
             }
         }
-    }, [isOn, characteristic]);
+    }, [isOn]);
 
-
-    const connectToDevice = () => {
-        setBluetoothStatus("Searching for device...");
-        navigator.bluetooth.requestDevice({
-            filters: [{ services: [DEVICE_SERVICE_UUID] }],
-            optionalServices: [DEVICE_SERVICE_UUID]
-        })
-            .then(device => {
-                setBluetoothStatus("Connecting to device...");
-                return device.gatt.connect();
-            })
-            .then(server => {
-                setBluetoothStatus("Discovering service...");
-                return server.getPrimaryService(DEVICE_SERVICE_UUID);
-            })
-            .then(service => {
-                setBluetoothStatus("Discovering characteristic...");
-                return service.getCharacteristic(DEVICE_CHARACTERISTIC_UUID);
-            })
-            .then(characteristic => {
[…]
