### 2024-03-15 4875f57
pull bluetooth into isolated component; bump mediapipe and dep load
