### 2024-03-21 #6392 @thejasvibr
> > Hi @arfon I ended being double-booked, by saying yes to two submissions in the pre-review phase and being finalised as a reviewer almost at the same time. Would already like to request some more time for this review (i.e. won't be able to make it in 2-4 weeks).
>
> No problem at all, and thanks for being a willing JOSS reviewer! What about trying to get your review in by mid-late April?

Hi @arfon , yes. Mid-late April sounds doable. Was out of action due to illness, and am getting back to shape.
