### 2024-03-21 mrilikecoding/StreamPoseML#15 @thejasvibr
Example notebooks with absolute paths to import package
Hi,
This is in relation to https://github.com/openjournals/joss-reviews/issues/6392

The example notebooks import the package using absolute paths. While this may be okay in the early development phases, it can mess up things further down the path - and also raises questions on the 'installability' of the package. Moreover, a usable package should be importable using the default pythonic ```import <packagename>``` command to keep things as simple as possible.

I checked at least two notebook and it seems like the absolute paths aren't really needed, and the code seems to work fine without them anyway. Would be good to clean up the absolute path import syntax.
