### 2024-03-21 mrilikecoding/StreamPoseML#16 @thejasvibr
Error in example notebook dataset_for_ui_example.ipynb
Hi,
This is in relation to https://github.com/openjournals/joss-reviews/issues/6392

I'm listing the first errors that popped up as I progressed through the cells in each notebook. These errors may perhaps also be a hardware issue as I'm running the notebooks on a laptop with non-CUDA GPU. Also please see #19

dataset_for_ui... : Cell 3
![image](https://github.com/mrilikecoding/StreamPoseML/assets/12443569/b338f67c-0239-42e2-8d19-a7a6aa39fdff)
...
```OSError: Cannot save file into a non-existent directory: '..\..\data\annotated_videos'```
