### 2024-03-21 mrilikecoding/StreamPoseML#17 @thejasvibr
Error in example notebook ```gradient_boosting_example```
Hi,
This is in relation to https://github.com/openjournals/joss-reviews/issues/6392

I'm listing the first errors that popped up as I progressed through the cells in each notebook. These errors may perhaps also be a hardware issue as I'm running the notebooks on a laptop with non-CUDA GPU. Also please see #19

gradient boosting... : Cell 5
![image](https://github.com/mrilikecoding/StreamPoseML/assets/12443569/5e9b4148-761d-4443-b2da-59085fdf5abd)
....intermediate error messages not shown...
```FileNotFoundError: [Errno 2] No such file or directory: '../../data/annotated_videos/pooled_angles_distances_full_examples.csv'```
