### 2024-03-21 mrilikecoding/StreamPoseML#18 @thejasvibr
Error in example notebook ```random_forest_example```
Hi,
This is in relation to https://github.com/openjournals/joss-reviews/issues/6392

I'm listing the first errors that popped up as I progressed through the cells in each notebook. These errors may perhaps also be a hardware issue as I'm running the notebooks on a laptop with non-CUDA GPU. Also please see #19

random_forest... : Cell 1
![image](https://github.com/mrilikecoding/StreamPoseML/assets/12443569/6122c447-56f4-4cd9-8afa-dba1ddbbc348)
