### 2024-03-21 mrilikecoding/StreamPoseML#19 @thejasvibr
Various errors with example notebooks
Hi,
This is in relation to https://github.com/openjournals/joss-reviews/issues/6392

It seems like the example notebooks haven't been tested on another system or environment with a 'clean' installation of the repo.
Each of the notebooks I tried to run threw a variety of errors.

I have raised multiple issues on the specifics, and this issue is a pointer on the need for an update + test on a clean installed system/environment,
