### 2024-03-24 mrilikecoding/StreamPoseML#19 @mrilikecoding
Thank you, I appreciate it! I'll start working through the issues. Cheers.
