+++ kubernetes.yml
@@ -1,0 +1,20 @@
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: streamposeml
+  labels:
+    app: streamposeml
+spec:
+  selector:
+    matchLabels:
+      app: streamposeml
+  template:
+    metadata:
+      labels:
+        app: streamposeml
+    spec:
+      containers:
+        - name: streamposeml
+          image: streamposeml-image
+          ports:
+            - containerPort: 5001
