### 2024-04-17 8894e12
allow model setting in memory from any file location
