+++ web_ui/tailwind.config.js
@@ -1,0 +1,17 @@
+/** @type {import('tailwindcss').Config} */
+export default {
+  content: [
+    "./index.html",
+    "./src/**/*.{js,ts,jsx,tsx}",
+  ],
+  theme: {
+    extend: {},
+  },
+  daisyui: {
+    themes: ["light", "dark", "retro", "lemonade"],
+  },
+  plugins: [
+    require("daisyui"),
+    require('@tailwindcss/typography'),
+  ],
+}
