+++ web_ui/src/helpers/api.js
@@ -1,35 +1,0 @@
-import axios from "axios";
-// import { getCookie } from "./utils";
-
-export default class Api {
-  constructor() {
-    this.api_token = null;
-    this.client = null;
-    this.api_url = process.env.REACT_APP_STREAM_POSE_ML_API_ENDPOINT;
-  }
-
-  init = () => {
-    // this.api_token = getCookie("ACCESS_TOKEN");
-
-    let headers = {
-      Accept: "application/json",
-    };
-
-    if (this.api_token) {
-      headers.Authorization = `Bearer ${this.api_token}`;
-    }
-
-    this.client = axios.create({
-      baseURL: this.api_url,
-      timeout: 31000,
-      headers: headers,
-    });
-
-    return this.client;
-  };
-
-  processVideo = (params) => {
-    console.log(this.api_url);
-    return this.init().get("/process_videos", { params: params });
-  };
-}
