+++ web_ui/Dockerfile
@@ -2,8 +2,12 @@ FROM node:18
 
 WORKDIR /usr/src/app/my-app
 
-COPY . ./
-
-RUN ls -al
+ADD package.json package.json
+ADD package-lock.json package-lock.json
 RUN npm install
+
+ADD . ./
+
 EXPOSE 3000
+ENTRYPOINT npm start
+
