### 2024-04-22 813ebac
add webui to tilt and remove unused mock bluetooth
