### 2024-04-26 ad41b12
add config to repo without need to copy example
