### 2024-04-26 mrilikecoding/StreamPoseML#19 @mrilikecoding
Greetings - thank you for your patience. I have consolidated the example notebooks into this single notebook which should be more generalized and hopefully work out of the box: https://github.com/mrilikecoding/StreamPoseML/blob/main/stream_pose_ml/notebooks/example_usage.ipynb. Please let me know if you have issues running it.

I will close the other notebook related issues and leave this one open pending your review. If all looks good I'll close out this issue. Thanks!
