### 2024-04-26 mrilikecoding/StreamPoseML#15 @mrilikecoding
I believe this issue is resolved in relation to https://github.com/mrilikecoding/StreamPoseML/issues/19. Closing this but please feel free to comment otherwise over there. Cheers.
