### 2024-04-29 e3c14d8
clean up some docs and deps
