+++ stream_pose_ml/requirements.txt
@@ -21,3 +21,4 @@ tqdm==4.65.0
 tslearn==0.6.2
 xgboost==1.7.4
 pytest
+mlflow
