### 2024-05-11 mrilikecoding/StreamPoseML#21 @thejasvibr
 unexpected keyword argument 'model_path'
In the example notebook - while saving the model for the web application I get this error. I've run the notebook as of the latest commit on ```main``` (e3c14d...)

![image](https://github.com/mrilikecoding/StreamPoseML/assets/12443569/ab9003cc-0458-468b-a192-ed0173efa76c)

```
TypeError: ModelBuilder.save_model_and_datasets() got an unexpected keyword argument 'model_path'
```

This is in relation to the JOSS review process (https://github.com/openjournals/joss-reviews/issues/6392)
