### 2024-05-11 mrilikecoding/StreamPoseML#22 @thejasvibr
'visual example' for step 7 of the schematic ('validate the model')
Hi,
The example notebook does a great job of showing the capabilities of the package.
One important thing may be to show a new user a visual example (e.g. overlay of body key-points on a previously unseen video) of what comes out from the trained model in the example notebook.
This is more likely to convince a first-time user who's not yet fully invested.

I see that there are already ROC curves etc. once the model is trained, but it might be useful to show the first-time casual user what the final outputs can actually look like?

Something optional, and completely left to you of course.

This is in relation to the JOSS review process (https://github.com/openjournals/joss-reviews/issues/6392)
