### 2024-05-11 mrilikecoding/StreamPoseML#23 @thejasvibr
identifiable person in the example data
Hi,
There's a person in the example data video set.
I'm honestly not sure how this is done with software packages and example data that come with it - but nonetheless the questions are - has this person given consent, and what are the isntitutional/community guidelines for this kind of video data?
