### 2024-05-17 mrilikecoding/StreamPoseML#23 @mrilikecoding
Hey there -- in this case that person is me, filmed for demo purposes, and all the example data is contrived. Do you think I need to be clear about that in the repo?
