### 2024-05-17 mrilikecoding/StreamPoseML#22 @mrilikecoding
Thanks, I'll give it some thought and follow up
