### 2024-07-07 #6392 @imcatta
Hi there,
I've finished the first round of my review. My comments/suggestions are related to the manuscript, as summarized here mrilikecoding/StreamPoseML#14
