### 2024-07-26 #6392 @thejasvibr
Hi @arfon ,
many apologies for my lack of response. I went back, and saw that I'd actually raised a couple of issues, but never followed them up some time ago.

I've filled out my review to the extent I can right now - two check-boxes are empty as of now, I explain below.
*human & animal data* : some of the example data has an identifiable person - what is JOSS's policy on this (the person is the package author)? (see [issue #23](https://github.com/mrilikecoding/StreamPoseML/issues/23))

*reproducibility* : I have run and verified that the CPU-runnable notebooks are working as expected. I do not have experience with Docker or the web-based aspects of this package, and will also not be able to do so because of unexpected personal commitments in the recent past and future. I have thus not tested or tried out these components.

sorry again for going silent - hope this review is still helpful.
