### 2024-08-27 mrilikecoding/StreamPoseML#19 @mrilikecoding
Closing per https://github.com/openjournals/joss-reviews/issues/6392#issuecomment-2252332707
