### 2024-08-27 mrilikecoding/StreamPoseML#23 @mrilikecoding
Hello, since I haven't heard if this is an issue (since the person in video is me shooting for demo purposes, closing this). If this is an issue still please let me know and I can come up with an alternate solution. Cheers
