### 2024-08-27 92c337d
work on new model transformer (WIP)
