+++ stream_pose_ml/stream_pose_ml/learning/sequence_transformer.py
@@ -16,6 +16,8 @@ class SequenceTransformer(ABC):
 from collections import defaultdict
 
 
+
+
 # TODO create concrete classes for different schemes
 class TenFrameFlatColumnAngleTransformer(SequenceTransformer):
     """This is the first transformer I'm writing here, so maybe a little more
@@ -69,3 +71,109 @@ class TenFrameFlatColumnAngleTransformer(SequenceTransformer):
         output_flattened = pd.json_normalize(data=output_dict)
         output_flattened_filtered = output_flattened.filter(columns)
         return (output_flattened_filtered, output_meta)
+
+
+class ThirtyFrameJointsTransformer(SequenceTransformer):
+    # joints.nose.x	joints.nose.y	joints.nose.z	joints.left_eye_inner.x	joints.left_eye_inner.y	...	joints.left_foot_index.z	joints.right_foot_index.x	joints.right_foot_index.y	joints.right_foot_index.z	joints.neck.x	joints.neck.y	joints.neck.z	joints.mid_hip.x	joints.mid_hip.y	joints.mid_hip.z
+
+    @staticmethod
+    def normalize_keypoints(group):
+        """ TODO extract to util """
+        # Choose a reference keypoint, for example, 'joints.nose.z'
+        reference_keypoint = 'joints.nose.z'
+        
+        # Calculate the initial distance using the first frame of this group
+        initial_distance = group[reference_keypoint].iloc[0]
+        
+        # Compute scale factors for each frame in the group
+        scale_factors = initial_distance / group[reference_keypoint]
+        
+        # Apply scaling to x and y coordinates
+        for col in group.columns:
+            if col.endswith('.x') or col.endswith('.y'):
+                group[col] = group[col] * scale_factors
+        
+        # Re-center the keypoints
+        centre_x = (group[[col for col in group.columns if col.endswith('.x')]].min(axis=1) + 
+                    group[[col for col in group.columns if col.endswith('.x')]].max(axis=1)) / 2
+        centre_y = (group[[col for col in group.columns if col.endswith('.y')]].min(axis=1) + 
+                    group[[col for col in group.columns if col.endswith('.y')]].max(axis=1)) / 2
+        
+        for col in group.columns:
+            if col.endswith('.x'):
+                group[col] = group[col] - centre_x
+            elif col.endswith('.y'):
+                group[col] = centre_y - group[col] 
+
+        
+        return group
+
+    def transform(self, data: any, columns: list) -> any:
+        joint_list = [
+            "nose",
+            "left_eye_inner",
+            "left_eye",
+            "left_eye_outer",
+            "right_eye_inner",
+            "right_eye",
+            "right_eye_outer",
+            "left_ear",
+            "right_ear",
+            "mouth_left",
+            "mouth_right",
+            "left_shoulder",
+            "right_shoulder",
+            "left_elbow",
+            "right_elbow",
+            "left_wrist",
+            "right_wrist",
+            "left_pinky",
+            "right_pinky",
+            "left_index",
+            "right_index",
+            "left_thumb",
+            "right_thumb",
+            "left_hip",
+            "right_hip",
+            "left_knee",
+            "right_knee",
+            "left_ankle",
+            "right_ankle",
+            "left_heel",
+            "right_heel",
+            "left_foot_index",
+            "right_foot_index",
+            "neck",
+            "mid_hip",
+        ]
+        df_data = []
+        column_names = [f"joints.{joint_name}.{dimension}" for joint_name in joint_list for dimension in ["x", "y", "z"]]
+        frames = data["frames"]
+        frame_joint_positions = []
+        dummy_columns = ['pid', 'video_index', 'is_step', 'weight_transfer_type', 'step_type']
+        column_names = dummy_columns + column_names
[…]
