### 2024-08-27 5482cc5
add default model path to model_builder
