### 2024-08-27 mrilikecoding/StreamPoseML#21 @mrilikecoding
@thejasvibr my apologies for the long delay (eventful summer). I believe I have fixed this bug now. Please try again at your convenience -- will close with your confirmation this works. Cheers
