### 2024-08-27 #6392 @mrilikecoding
Hi @arfon, my apologies for the delays. It's been an eventful summer on my end.

Re: the human & animal data [issue #23](https://github.com/mrilikecoding/StreamPoseML/issues/23), I closed because I am the person in the video and I did it for demo purposes. Let me know if this is an issue still.

I fixed another bug[ #21](https://github.com/mrilikecoding/StreamPoseML/issues/21), and will close with confirmation.

For [#22](https://github.com/mrilikecoding/StreamPoseML/issues/22), I may defer that one since it is an optional suggestion but will do it if you all prefer. Let me know.

Otherwise that leaves [issue #14](https://github.com/mrilikecoding/StreamPoseML/issues/14) which work on.

Thanks
