### 2024-09-14 c7cfdf1
add mlflow model logging to model builder
