### 2024-09-14 70ac007
hyphenate model type for file saving convenience
