### 2024-09-24 142f059
add mlflow service that loads model via passed path
