+++ mlflow/app.py
@@ -1,0 +1,76 @@
+from flask import Flask, request, jsonify
+import subprocess
+import threading
+import os
+import signal
+import time
+
+app = Flask(__name__)
+
+# Global variable to keep track of the model server process
+model_server_process = None
+
+@app.route('/status', methods=['GET'])
+def status():
+    return jsonify({'status': 'running', 'message': 'MLflow model server is up and running.'}), 200
+
+@app.route('/ping', methods=['POST'])
+def ping():
+    if model_server_process:
+        return jsonify({'status': 'success', 'message': 'Model server is running.'}), 200
+    else:
+        return jsonify({'status': 'error', 'message': 'Model server not running.'}), 500
+
+@app.route('/load_model', methods=['POST'])
+def load_model():
+    global model_server_process
+
+    # Get the model path from the request
+    data = request.get_json()
+    model_path = data.get('model_path')
+
+    if not model_path or not os.path.exists(model_path):
+        return jsonify({'status': 'error', 'message': 'Invalid model path.'}), 400
+
+    # Terminate existing model server if running
+    if model_server_process:
+        os.killpg(os.getpgid(model_server_process.pid), signal.SIGTERM)
+        model_server_process = None
+        time.sleep(2)  # Wait for process to terminate
+
+    # Start the MLflow model serving process
+    command = [
+        'mlflow', 'models', 'serve',
+        '-m', model_path,
+        '-p', '1234',  # Use a different port
+        '-h', '0.0.0.0'
+    ]
+
+    model_server_process = subprocess.Popen(
+        command,
+        stdout=subprocess.PIPE,
+        stderr=subprocess.PIPE,
+        preexec_fn=os.setsid  # Start the process in a new process group
+    )
+
+    return jsonify({'status': 'success', 'message': 'Model server started.'}), 200
+
+@app.route('/invocations', methods=['POST'])
+def invocations():
+    if not model_server_process:
+        return jsonify({'status': 'error', 'message': 'Model server not running.'}), 500
+
+    # Forward the request to the MLflow model server
+    import requests
+    try:
+        response = requests.post(
+            'http://localhost:1234/invocations',
+            data=request.data,
+            headers={'Content-Type': request.headers.get('Content-Type')}
+        )
+        return (response.content, response.status_code, response.headers.items())
+    except requests.exceptions.ConnectionError:
+        return jsonify({'status': 'error', 'message': 'Failed to connect to model server.'}), 500
+
+if __name__ == '__main__':
+    app.run(host='0.0.0.0', port=5002, debug=True)
