### 2024-09-24 c7f256d
add mlflow service that loads model via passed path
