### 2024-09-25 d558971
allow web_ui to set mlflow model
