### 2024-09-25 6a86020
allow web_ui to set mlflow model
