+++ mlflow/Dockerfile
@@ -1,10 +1,10 @@
-# Use Python base image
-FROM python:3.10-slim
+# Use a base image with Conda installed
+FROM continuumio/miniconda3
 
-# Install necessary packages
+# Install MLflow and any additional base dependencies
 RUN pip install --no-cache-dir mlflow flask requests
 
-# Expose port 5000 for the Flask API
+# Expose port 5002 for the Flask API
 EXPOSE 5002
 
 # Copy the application code
