### 2024-09-26 c4a6104
update debug settings
