### 2024-09-26 49aeea8
load model from zip via ui
