### 2024-09-30 b1973a0
log model signature for mlflow model
