### 2024-10-09 52214d1
rough in input data  transformation
