+++ stream_pose_ml/api/app.py
@@ -2,7 +2,7 @@ import time
 import os
 from pathlib import Path
 import logging
-
+import json
 
 from flask import Flask, request, jsonify
 
@@ -30,7 +30,7 @@ trained_model = trained_model.TrainedModel()
 
 
 ### Set the pose estimation client ###
-# dummy_client=True here indicates that we don't need mediapipe loaded, 
+# dummy_client=True here indicates that we don't need mediapipe loaded,
 # we just want to use methods on the class
 mpc = mediapipe_client.MediaPipeClient(dummy_client=True)
 
@@ -60,25 +60,28 @@ app = Flask(__name__)
 app.config["SECRET_KEY"] = "secret!"
 
 # Tmp file upload
-UPLOAD_FOLDER = 'tmp'
-ALLOWED_EXTENSIONS = {'zip', 'tar.gz', 'tar', 'pickle', 'joblib', 'model'}
+UPLOAD_FOLDER = "tmp"
+ALLOWED_EXTENSIONS = {"zip", "tar.gz", "tar", "pickle", "joblib", "model"}
 
-app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
+app.config["UPLOAD_FOLDER"] = UPLOAD_FOLDER
 Path(UPLOAD_FOLDER).mkdir(parents=True, exist_ok=True)
 
 
 # TODO - make env dependent from config
 whitelist = [
     "http://localhost:3000",
+    "http://web_ui:3000",
     "http://localhost:5001",
+    "http://stream_pose_ml:5001",
     "https://cdn.jsdelivr.net",
 ]
 CORS(app, origins=whitelist)
 # CORS(app, origins="*")
 
+
 def allowed_file(filename):
-    return '.' in filename and \
-           filename.rsplit('.', 1)[1].lower() in ALLOWED_EXTENSIONS
+    return "." in filename and filename.rsplit(".", 1)[1].lower() in ALLOWED_EXTENSIONS
+
 
 @app.route("/")
 def status():
@@ -88,14 +91,14 @@ def status():
 ### Application Routes ###
 @app.route("/set_model", methods=["POST"])
 def set_model():
-    if 'file' not in request.files:
+    if "file" not in request.files:
         return jsonify({"result": "No file part"}), 400
-    file = request.files['file']
+    file = request.files["file"]
     if file.filename == "":
         return jsonify({"result": "No selected file"}), 400
     if file and allowed_file(file.filename):
         filename = secure_filename(file.filename)
-        model_path = os.path.join(app.config['UPLOAD_FOLDER'], filename)
+        model_path = os.path.join(app.config["UPLOAD_FOLDER"], filename)
         file.save(model_path)
 
         # Extract the archive
@@ -104,26 +107,39 @@ def set_model():
         )
 
         # Handle different archive formats
-        if filename.endswith('.zip'):
-            with zipfile.ZipFile(model_path, 'r') as zip_ref:
+        if filename.endswith(".zip"):
+            with zipfile.ZipFile(model_path, "r") as zip_ref:
                 zip_ref.extractall(extract_to)
             model_path = extract_to
-        elif filename.endswith('.tar.gz') or filename.endswith('.tar'):
+        elif filename.endswith(".tar.gz") or filename.endswith(".tar"):
             with tarfile.open(model_path, "r:*") as tar_ref:
                 tar_ref.extractall(extract_to)
             model_path = extract_to
 
         # Send a request to mlflow to load the model
-        input_example = None # TODO parse the input example
         mlflow_response = load_model_in_mlflow(model_path)
         if mlflow_response:
+            # Load input_example.json if it exists
+            input_example_path = os.path.join(model_path, "input_example.json")
+            input_example = None
+            if os.path.exists(input_example_path):
[…]
