### 2024-10-30 865dce4
get transformer working and send to predict fn
