+++ mlflow/app.py
@@ -1,5 +1,6 @@
 from flask import Flask, request, jsonify
 import subprocess
+import requests
 import threading
 import os
 import signal
@@ -9,34 +10,45 @@ import logging
 app = Flask(__name__)
 
 # Configure logging
-logging.basicConfig(level=logging.INFO, format='%(asctime)s %(message)s')
+logging.basicConfig(level=logging.INFO, format="%(asctime)s %(message)s")
 logger = logging.getLogger(__name__)
 
 # Global variable to keep track of the model server process
 model_server_process = None
 
-@app.route('/status', methods=['GET'])
-def status():
-    return jsonify({'status': 'running', 'message': 'MLflow model server is up and running.'}), 200
 
-@app.route('/ping', methods=['POST'])
+@app.route("/status", methods=["GET"])
+def status():
+    return (
+        jsonify(
+            {"status": "running", "message": "MLflow model server is up and running."}
+        ),
+        200,
+    )
+
+
+@app.route("/ping", methods=["POST"])
 def ping():
     if model_server_process:
-        return jsonify({'status': 'success', 'message': 'Model server is running.'}), 200
+        return (
+            jsonify({"status": "success", "message": "Model server is running."}),
+            200,
+        )
     else:
-        return jsonify({'status': 'error', 'message': 'Model server not running.'}), 500
+        return jsonify({"status": "error", "message": "Model server not running."}), 500
 
-@app.route('/load_model', methods=['POST'])
+
+@app.route("/load_model", methods=["POST"])
 def load_model():
     global model_server_process
 
     # Get the model path from the request
     data = request.get_json()
-    model_path = data.get('model_path')
+    model_path = data.get("model_path")
 
     if not model_path or not os.path.exists(model_path):
         print("invalid model path")
-        return jsonify({'status': 'error', 'message': 'Invalid model path.'}), 400
+        return jsonify({"status": "error", "message": "Invalid model path."}), 400
 
     # Terminate existing model server if running
     if model_server_process:
@@ -46,27 +58,30 @@ def load_model():
 
     # Start the MLflow model serving process
     command = [
-        'mlflow', 'models', 'serve',
-        '-m', model_path,
-        '-p', '1234',  # Use a different port
-        '-h', '0.0.0.0'
+        "mlflow",
+        "models",
+        "serve",
+        "-m",
+        model_path,
+        "-p",
+        "1234",  # Use a different port
+        "-h",
+        "0.0.0.0",
     ]
 
     model_server_process = subprocess.Popen(
         command,
         stdout=subprocess.PIPE,
         stderr=subprocess.PIPE,
-        preexec_fn=os.setsid  # Start the process in a new process group
+        preexec_fn=os.setsid,  # Start the process in a new process group
     )
 
-        # Start threads to read subprocess output
+    # Start threads to read subprocess output
     stdout_thread = threading.Thread(
-        target=read_subprocess_output, 
-        args=(model_server_process.stdout, logger.info)
+        target=read_subprocess_output, args=(model_server_process.stdout, logger.info)
     )
     stderr_thread = threading.Thread(
-        target=read_subprocess_output, 
-        args=(model_server_process.stderr, logger.error)
+        target=read_subprocess_output, args=(model_server_process.stderr, logger.error)
     )
     stdout_thread.start()
     stderr_thread.start()
@@ -75,13 +90,14 @@ def load_model():
     monitor_thread = threading.Thread(target=monitor_model_server)
     monitor_thread.start()
 
+    return jsonify({"status": "success", "message": "Model server started."}), 200
 
-    return jsonify({'status': 'success', 'message': 'Model server started.'}), 200
 
 def read_subprocess_output(pipe, logger_method):
-    for line in iter(pipe.readline, b''):
[…]
