### 2024-10-30 2714b3e
get transformer working and send to predict fn
