### 2024-11-11 16e6238
Add teachable machine ref to paper.bib
