+++ paper.bib
@@ -77,3 +77,10 @@
   year={2021},
   publisher={MDPI}
 }
+@misc{teachablemachine,
+  title={Teachable Machine: A web-based tool for creating machine learning models},
+  author={{Google}},
+  howpublished={\url{https://teachablemachine.withgoogle.com/}},
+  year={2024},
+  note={Accessed: 2024-11-11}
+}
