### 2024-11-12 mrilikecoding/StreamPoseML#22 @mrilikecoding
I'm going to defer on this for now in the interest of time. If reviewers feel strongly happy to re=approach. Thanks
