### 2024-11-12 mrilikecoding/StreamPoseML#14 @mrilikecoding
Hello @imcatta, thanks, and apologies for the very long delay. I believe I have addressed this now. I ended up removing the figure for a few reasons in favor of a text list. I added the reference to teachable machine. Additionally, I noticed the paper action was deprecated so updated that and the paper seems to be compiling now.
