### 2024-11-12 #6392 @mrilikecoding
Hello @arfon, thanks for your patience. I have closed out mrilikecoding/StreamPoseML#14 now. I believe with that I have addressed all the issues related to the JOSS submission here. Please let me know if there is anything else, I'm doing some other work on the package this week and next. Cheers.
