### 2024-11-12 dc95335
format data for model server with dummy defaults for successful prediction
