### 2024-11-12 4a9f6f3
format data for model server with dummy defaults for successful prediction
