### 2024-11-15 45a32c8
throttle prediction based on frame window size an overlap params
