### 2024-11-15 9cf51dc
throttle prediction based on frame window size an overlap params
