+++ stream_pose_ml/stream_pose_ml/ml_flow_client.py
@@ -1,3 +1,4 @@
+from threading import current_thread
 import time
 import typing
 
@@ -17,14 +18,16 @@ if typing.TYPE_CHECKING:
 class MLFlowClient:
     def __init__(
         self,
-        frame_window: int = 25,
         mediapipe_client_instance: type["MediaPipeClient"] | None = None,
         trained_model: type["TrainedModel"] | None = None,
         data_transformer: type["SequenceTransformer"] | None = None,
         predict_fn: typing.Callable | None = None,
         input_example: dict = {"columns": []},
+        frame_window: int = 30,
+        frame_overlap: int = 5,
     ):
         self.frame_window = frame_window
+        self.frame_overlap = frame_overlap
         self.model = trained_model
         self.transformer = data_transformer
         self.mpc = mediapipe_client_instance
@@ -33,11 +36,19 @@ class MLFlowClient:
         self.predict_fn = predict_fn
         self.input_example = input_example
         self.input_example_columns = input_example["columns"]
+        self.update_frame_frequency = self.frame_window - self.frame_overlap
+        self.counter = 0
+        self.prediction_processing_time = None
+        self.last_prediction_timestamp = 0
 
     def run_keypoint_pipeline(self, keypoints):
         current_frames = self.update_frame_data_from_js_client_keypoints(keypoints)
-        if len(current_frames) == self.frame_window:
-            # TODO this could probably be throttled based on configured predict frequency
+        self.counter += 1
+        if (
+            len(current_frames) == self.frame_window
+            and self.counter >= self.update_frame_frequency
+        ):
+            self.counter = 0
             sequence = BlazePoseSequence(
                 name=f"sequence-{time.time_ns()}",
                 sequence=list(current_frames),
@@ -52,9 +63,15 @@ class MLFlowClient:
                 return
 
             # TODO enforce signature of predict_fn, this is brittle
+            start_time = time.time()
             prediction = self.predict_fn(json_data_payload=data)["predictions"][0]
+            current_time = time.time()
+            speed = current_time - start_time
+            self.prediction_processing_time = speed
+            self.last_prediction_timestamp = current_time
             self.current_classification = bool(prediction)
-        return True
+            return True
+        return False
 
     def update_frame_data_from_js_client_keypoints(self, keypoint_results):
         frame_data = {
@@ -71,4 +88,5 @@ class MLFlowClient:
             )
 
             self.frames.append(frame_data)
+
         return self.frames
