### 2024-11-20 c72c964
update build scripts
