+++ pyproject.toml
@@ -6,21 +6,21 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "stream_pose_ml"
-version = "0.1.5" # placeholder, will be overwritten by setuptools_scm
-authors = [{name = "Nate Green", email = "nate@nate.green"}]
+version = "0.2.0"
+authors = [{ name = "Nate Green", email = "nate@nate.green" }]
 description = "A toolkit for realtime video classification tasks."
 readme = "README.md"
 dynamic = ["dependencies"]
-license = {file = "LICENSE"}
+license = { file = "LICENSE" }
 classifiers = [
-    "Development Status :: 3 - Alpha",
-    "Intended Audience :: Developers",
-    "Topic :: Software Development :: Libraries :: Python Modules",
-    "Topic :: Scientific/Engineering :: Artificial Intelligence",
-    "License :: OSI Approved :: MIT License",
-    "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.10",
-    "Programming Language :: Python :: 3.11"
+  "Development Status :: 3 - Alpha",
+  "Intended Audience :: Developers",
+  "Topic :: Software Development :: Libraries :: Python Modules",
+  "Topic :: Scientific/Engineering :: Artificial Intelligence",
+  "License :: OSI Approved :: MIT License",
+  "Programming Language :: Python :: 3",
+  "Programming Language :: Python :: 3.10",
+  "Programming Language :: Python :: 3.11",
 ]
 
 [project.urls]
