### 2024-12-07 #6392 @mrilikecoding
@arfon sounds good. I've done a final pass and generated a new release. The automatic update on Zenodo didn't work for some reason but I was able to create the archive manually. Here is the new DOI: https://doi.org/10.5281/zenodo.14298482 - let me know if anything looks amiss. Thanks!
