### 2024-12-27 #6392 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
