### 2024-12-27 #6392 @arfon
@mrilikecoding – could you please merge this PR which makes a few formatting fixes to your paper? https://github.com/mrilikecoding/StreamPoseML/pull/25
